requirements.txt
sqlmodel==0.0.14
fastapi==0.109.0
python-dotenv==1.0.0
